feat: disconnect role from entity pop up
